app/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
npm==0.1.1
optional-django==0.1.0
Pillow==9.3.0
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
twilio==8.0.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
gunicorn==20.1.0
djangorestframework-simplejwt==5.2.2
django-jazzmin


